packages/renouveau-schema/pyproject.toml
[project]
name = "renouveau-schema"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Alexis Beaulieu", email = "alexisbeaulieu97@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
]

[project.scripts]
renouveau-schema = "renouveau_schema:main"

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"
